# frontend/openapi/auth.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Message": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "Message",
        "type": "object"
      },
      "RefreshTokenPublic": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ip_address": {
            "maxLength": 45,
            "title": "Ip Address",
            "type": "string"
          },
          "last_used_at": {
            "format": "date-time",
            "title": "Last Used At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "expires_at",
          "ip_address",
          "id",
          "user_id"
        ],
        "title": "RefreshTokenPublic",
        "type": "object"
      },
      "RefreshTokensPublic": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "refresh_tokens": {
            "items": {
              "$ref": "#/components/schemas/RefreshTokenPublic"
            },
            "title": "Refresh Tokens",
            "type": "array"
          }
        },
        "required": [
          "refresh_tokens",
          "count"
        ],
        "title": "RefreshTokensPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n    A REST API service for user authentication.\n\n    This REST API:\n    - Allows users to register and login.\n    - Generates access tokens and refresh tokens.\n    - Allows users to refresh their access tokens.\n    - Allows users to logout.\n    - Allows users to view and revoke their refresh tokens.\n    ",
    "title": "Microservices [auth]",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tokens/me": {
      "get": {
        "description": "Get the current user refresh tokens.\n\nArgs:\n    session (AsyncSession): The database session.\n    user_token (TokenData): The user's token data.\n\nReturns:\n    RefreshTokensPublic: The current user sessions.",
        "operationId": "get_user_refresh_tokens_tokens_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokensPublic"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get User Refresh Tokens",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/me/{token_id}": {
      "delete": {
        "description": "Revoke a user refresh token.\n\nArgs:\n    session (AsyncSession): The database session.\n    user_token (TokenData): The user's token data.\n    token_id (UUID): The refresh token id.\n\nReturns:\n    Message: The response message.",
        "operationId": "revoke_user_refresh_token_tokens_me__token_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Revoke User Refresh Token",
        "tags": [
          "tokens"
        ]
      }
    },
...
    "/verification/email": {
      "get": {
        "description": "Send verification email.\n\nArgs:\n    session (AsyncSession): The database session.\n    email (str): The email address to send the verification email to.\n\nReturns:\n    Message: The response message.",
        "operationId": "send_verification_email_verification_email_get",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "title": "Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Verification Email",
        "tags": [
          "verification"
        ]
      }
    },
    "/verification/email/{token}": {
      "get": {
        "description": "Verify email address.\n\nArgs:\n    session (AsyncSession): The database session.\n    token (str): The verification token.\n\nReturns:\n    Message: The response message.",
        "operationId": "verify_email_verification_email__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Email",
        "tags": [
          "verification"
        ]
      }
    }
  }
}